+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add a `DashBoard` React based view
 - Add API endpoint to search & autocomplete licences
+- Add licences-based filters on course search
 
 ### Changed